requirements.txt
streamlit>=1.37
python-dotenv>=1.0
langchain>=0.2.12
langchain-community>=0.2.10
langchain-groq>=0.1.5
faiss-cpu>=1.8.0
pypdf>=4.3.1
sentence-transformers>=3.0.1